third_party/blink/web_tests/fast/events/pointerevents/mouse-pointer-capture-in-iframe.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/resources/testdriver-actions.js"></script>
<style>
iframe {
  width: 300px;
  height: 300px;
  top: 100px;
  left: 100px;
  border: 0;
  position: absolute;
  background: green;
}
#outerFrame {
  width: 500px;
  height: 500px;
  background: blue;
}
</style>
<body id="outerFrame body">
<div id='outerFrame'>
<iframe id='innerFrameElement' srcdoc="
  <html id='innerFrameDocument'>
  <body id='innerFrame' style='height:500px; width: 500px; padding: 0; margin: 0;'>
    <script>
      top.document.testEventList.forEach(function(eventName) {
        document.addEventListener(eventName, top.handleEvent);
      });
    </script>
  </body>
  </html>">
</iframe>
</div>
</body>
<script>
var receivedEventList = [];
function handleEvent(event) {
  receivedEventList.push(event.target.id + ' received ' + event.type);

  if (event.type == 'pointerdown') {
    if (document.preventDefaultOnPointerDown) {
      event.preventDefault();
    }
  }
};

document.testEventList = ['pointerup', 'pointerdown', 'pointermove', 'gotpointercapture', 'lostpointercapture'];
document.testEventList.forEach(function(eventName) {
  document.getElementById('outerFrame').addEventListener(eventName, handleEvent);
});

document.preventDefaultOnPointerDown = false;

promise_test (async() => {
    receivedEventList = [];
    expectedEventList = ["outerFrame received pointermove",
                         "outerFrame received pointerdown",
                         "outerFrame received pointermove",
                         "innerFrame received pointermove",
                         "outerFrame received pointermove",
                         "innerFrame received pointermove",
                         "innerFrame received pointerup"];
    await new test_driver.Actions()
                         .pointerMove(50, 50)
                         .pointerDown()
                         .pointerMove(60, 60)
                         .pointerMove(200, 200)
                         .pointerMove(250, 450)
                         .pointerMove(200, 200)
                         .pointerUp()
                         .send();

    assert_array_equals(receivedEventList, expectedEventList);
}, "Mouse down at outer frame and move to inner frame, without set pointer capture.");


promise_test (async() => {
    receivedEventList = [];
    expectedEventList = ["innerFrame received pointermove",
                         "innerFrame received pointerdown",
                         "innerFrame received pointermove",
                         "innerFrameDocument received pointermove",
                         "innerFrameDocument received pointermove"];
    await new test_driver.Actions()
                         .pointerMove(200, 200)
                         .pointerDown()
                         .pointerMove(250, 250)
                         .pointerMove(50, 50)
                         .pointerMove(450, 450)
                         .send();
    assert_array_equals(receivedEventList, expectedEventList);
}, "Test the frame capture behavior: Mouse down at inner frame and move to outer frame without set pointer capture.");


promise_test (async() => {
    document.preventDefaultOnPointerDown = true;
    receivedEventList = [];
    expectedEventList = ["innerFrame received pointermove",
                         "innerFrame received pointerdown",
                         "innerFrame received pointermove",
                         "outerFrame received pointermove",
                         "outerFrame received pointermove",
                         "outerFrame received pointerup"];
    await new test_driver.Actions()
                         .pointerMove(250, 250)
                         .pointerDown()
                         .pointerMove(200, 200)
                         .pointerMove(50, 50)
                         .pointerMove(450, 450)
                         .pointerUp()
                         .send();
    assert_array_equals(receivedEventList, expectedEventList);
    document.preventDefaultOnPointerDown = false;
}, "Mouse down with preventDefault at inner frame and move to outer frame.");
</script>
